+++ tests/input_capture.rs
@@ -40,17 +40,6 @@ fn capture_app() -> App {
 }
 
 #[test]
-fn app_update_sends_app_exit() {
-    let mut app = capture_app();
-
-    app.update();
-    let timestamped_input = app.world.resource::<TimestampedInputs>();
-    assert_eq!(timestamped_input.len(), 1);
-    let timestamped_event = timestamped_input.events.first().unwrap();
-    assert_eq!(timestamped_event.input_event, InputEvent::AppExit);
-}
-
-#[test]
 fn capture_sent_events() {
     let mut app = capture_app();
 
@@ -60,7 +49,7 @@ fn capture_sent_events() {
 
     app.update();
     let timestamped_input = app.world.resource::<TimestampedInputs>();
-    assert_eq!(timestamped_input.len(), 3);
+    assert_eq!(timestamped_input.len(), 2);
 }
 
 #[test]
@@ -82,13 +71,11 @@ fn identity_of_sent_events() {
 
     let first_event: TimestampedInputEvent = iterator.next().unwrap();
     let second_event: TimestampedInputEvent = iterator.next().unwrap();
-    let third_event: TimestampedInputEvent = iterator.next().unwrap();
 
     // Unfortunately these input types don't impl PartialEq :(
     assert!(matches!(first_event.input_event, InputEvent::Keyboard(_)));
-    assert!(matches!(second_event.input_event, InputEvent::AppExit));
     assert!(matches!(
-        third_event.input_event,
+        second_event.input_event,
         InputEvent::MouseButton(_)
     ));
 }
@@ -109,16 +96,13 @@ fn framecount_of_sent_events() {
     let mut timestamped_input = app.world.resource_mut::<TimestampedInputs>();
     let mut iterator = timestamped_input.iter_all().into_iter();
 
-    let first_event: TimestampedInputEvent = iterator.next().unwrap();
-    let second_event: TimestampedInputEvent = iterator.next().unwrap();
-    let third_event: TimestampedInputEvent = iterator.next().unwrap();
+    let first_event: TimestampedInputEvent = iterator.next().expect("Keyboard event failed.");
+    let second_event: TimestampedInputEvent = iterator.next().expect("Mouse event failed.");
 
     // The frame count is recorded based on the frame it is read,
     // which counts up immediately
     assert_eq!(first_event.frame, FrameCount(1));
-    // The app exit event
-    assert_eq!(second_event.frame, FrameCount(1));
-    assert_eq!(third_event.frame, FrameCount(2));
+    assert_eq!(second_event.frame, FrameCount(2));
 }
 
 #[test]
@@ -133,7 +117,7 @@ fn toggle_input_capture() {
 
     // Inputs are captured while input capturing is enabled by default
     let timestamped_input = app.world.resource::<TimestampedInputs>();
-    assert_eq!(timestamped_input.len(), 3);
+    assert_eq!(timestamped_input.len(), 2);
 
     // Disabling input capture
     let mut input_modes_captured = app.world.resource_mut::<InputModesCaptured>();
@@ -146,9 +130,8 @@ fn toggle_input_capture() {
     app.update();
 
     // Inputs are not captured while input capturing is disabled
-    // Note that each app.update() always sends an `AppExit` event
     let timestamped_input = app.world.resource::<TimestampedInputs>();
-    assert_eq!(timestamped_input.len(), 4);
+    assert_eq!(timestamped_input.len(), 2);
 
     // Partially re-enabling input capture
     let mut input_modes_captured = app.world.resource_mut::<InputModesCaptured>();
@@ -168,5 +151,5 @@ fn toggle_input_capture() {
 
     // Only the keyboard events (and app exit events) were captured
     let timestamped_input = app.world.resource::<TimestampedInputs>();
-    assert_eq!(timestamped_input.len(), 6);
+    assert_eq!(timestamped_input.len(), 3);
 }