+++ RELEASES.md
@@ -4,6 +4,8 @@
 
 Migrated to `bevy 0.10`.
 
+Note that `App::update` no longer sends an `AppExit` event: this may affect your tests!
+
 ## Version 0.1
 
 ### Enhancements